Create README.md
A repository filled with CodeWars challenges (kata). I currently solve one solution a day to increase my knowledge, sharpen my skills, and engage with the community. The biggest lesson I've learned is that the brain is a tool that we must exercise everyday. Each day is an opportunity to grow and evolve!